chore: Add geocoding service
This commit adds a new file, apiGeocoding.ts, which contains a function for retrieving address information based on latitude and longitude coordinates. The function makes use of the BigDataCloud API to perform reverse geocoding. If the API request fails, an error is thrown. The retrieved data is returned as the result of the function.